# app/controller/redash.go
package controller

import (
    "git.puls.ru/devops1/sre/devops-api/model"
	"git.puls.ru/devops1/sre/devops-api/types"
	"github.com/gin-gonic/gin"
    _ "git.puls.ru/devops1/sre/devops-api/docs" // docs is generated by Swag CLI, you have to import it.
  )



// @Summary Send message to "PULS Markirovka Alerts" telegram chat
// @Description Send message "PULS Markirovka Alerts" to telegram chat
// @Tags redash
// @Accept json
// @Produce json
// @Param message body types.Alert true "Send message"
// @Success 200 {array} types.Message
// @Security ApiKeyAuth
// @Router /redash/markirovka [post]
func (c *Controller) RedashMarkirovka(ctx *gin.Context) {
	model.AlertRedash(ctx, types.ChatIdMarkirovka, types.BotTokenMarkirovka)
}

// @Summary Send message to "PULS DevOps Monitor" telegram chat
// @Description Send message "PULS DevOps Monitor" to telegram chat
// @Tags redash
// @Accept json
// @Produce json
// @Param message body types.Alert true "Send message"
// @Success 200 {array} types.Message
// @Failure 400,404 {array} model.HTTPError
// @Failure 500 {array} model.HTTPError
// @Security ApiKeyAuth
// @Router /redash/devops [post]
func (c *Controller) RedashDevops(ctx *gin.Context) {
	model.AlertRedash(ctx, types.ChatIdDevops, types.BotTokenDevops)
}

// @Summary Send message to "GRVS support alerts" telegram chat
// @Description Send message "GRVS support alerts" to telegram chat
// @Tags redash
// @Accept json
// @Produce json
// @Param message body types.Alert true "Send message"
// @Success 200 {array} types.Message
// @Failure 400,404 {array} model.HTTPError
// @Failure 500 {array} model.HTTPError
// @Security ApiKeyAuth
// @Router /redash/ovs-support [post]
func (c *Controller) RedashOVSSupport(ctx *gin.Context) {
	model.AlertRedash(ctx, types.ChatIdOVSSupport, types.BotTokenDevops)
}